\name{Paneldata.formula}
\alias{Paneldata.formula}
\title{formula}
\usage{
\method{Paneldata}{formula}(formula, data = list(), n, t, model = c("fe",
  "re"), ...)
}
\arguments{
  \item{formula}{PIB~INF+TIR}

  \item{data}{the dataframe}

  \item{n}{the number of section}

  \item{t}{the time per section}

  \item{model}{if fixed==>"fe" if randon==>"re"}

  \item{...}{not used}
}
\description{
formula
}

